model/requirements.txt
python-3.11.9

pandas
google-generativeai
python-dotenv
rich
pyperclip

langchain>=0.2.2
langchain_core
langchain-core
langchain_community


python-dotenv
langchain-google-genai
google-generativeai
pydantic


pdf2image
pytesseract
pypdf
pydantic

torch
sentence_transformers
faiss-cpu


langchain-huggingface>=0.0.1
sentence-transformers>=2.2.2

fastapi 
uvicorn

